mcp/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
Authlib==1.6.4
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.1
cyclopts==3.24.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.12.3
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.14.0
mdurl==0.1.2
more-itertools==10.8.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
parse==1.20.2
pathable==0.4.4
pg8000==1.30.5
psycopg2-binary==2.9.6
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.27.1
scramp==1.4.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.1
Werkzeug==3.1.1
